refactor: Fix kitty protoccol support
