+++ cli/src/ui/utils/terminalCapabilities.ts
@@ -4,40 +4,6 @@
  */
 
 /**
- * Query terminal for Kitty keyboard protocol support
- * Returns a promise that resolves to true if supported, false otherwise
- */
-export async function detectKittyProtocolSupport(): Promise<boolean> {
-	return new Promise((resolve) => {
-		// Set a timeout in case terminal doesn't respond
-		const timeout = setTimeout(() => {
-			resolve(false)
-		}, 100) // 100ms timeout
-
-		// Listen for terminal response
-		const onData = (data: Buffer) => {
-			const response = data.toString()
-
-			// Check for Kitty protocol query response
-			// Response format: ESC [ ? <flags> u
-			if (response.match(/\x1b\[\?(\d+)u/)) {
-				clearTimeout(timeout)
-				process.stdin.removeListener("data", onData)
-				resolve(true)
-				return
-			}
-		}
-
-		// Setup listener
-		process.stdin.on("data", onData)
-
-		// Query terminal for keyboard protocol support
-		// CSI ? u - Query keyboard protocol
-		process.stdout.write("\x1b[?u")
-	})
-}
-
-/**
  * Detect terminal type from environment variables
  */
 export function detectTerminalType(): string {
@@ -58,12 +24,11 @@ export function detectTerminalType(): string {
 /**
  * Check if terminal is likely to support Kitty protocol based on type
  */
-export function isLikelyKittyProtocolSupported(): boolean {
+export function detectKittyProtocolSupport(): boolean {
 	const termType = detectTerminalType()
 
 	// Known terminals with Kitty protocol support
 	const supportedTerminals = ["kitty", "wezterm", "alacritty"]
-
 	return supportedTerminals.includes(termType)
 }
 
@@ -71,14 +36,9 @@ export function isLikelyKittyProtocolSupported(): boolean {
  * Auto-detect and enable Kitty protocol if supported
  * Returns true if enabled, false otherwise
  */
-export async function autoEnableKittyProtocol(): Promise<boolean> {
-	// First check if likely supported based on terminal type
-	if (!isLikelyKittyProtocolSupported()) {
-		// Still try to detect, but less likely to succeed
-	}
-
+export function autoEnableKittyProtocol(): boolean {
 	// Query terminal for actual support
-	const isSupported = await detectKittyProtocolSupport()
+	const isSupported = detectKittyProtocolSupport()
 
 	if (isSupported) {
 		// Enable Kitty keyboard protocol
